feat: Refactor data structs and functions
Moving signing operations into eth-bridge and primitive modules
