+++ pallets/avn/src/lib.rs
@@ -34,8 +34,7 @@ use sp_avn_common::{
     bounds::{MaximumValidatorsBound, ProcessingBatchBound},
     event_types::{EthEvent, EthEventId, Validator},
     ocw_lock::{self as OcwLock, OcwStorageError},
-    recover_public_key_from_ecdsa_signature, DEFAULT_EXTERNAL_SERVICE_PORT_NUMBER,
-    EXTERNAL_SERVICE_PORT_NUMBER_KEY,
+    DEFAULT_EXTERNAL_SERVICE_PORT_NUMBER, EXTERNAL_SERVICE_PORT_NUMBER_KEY,
 };
 use sp_core::{ecdsa, H160};
 use sp_runtime::{
@@ -430,42 +429,6 @@ impl<T: Config> Pallet<T> {
             signature_valid
         );
         return signature_valid
-    }
-
-    pub fn eth_signature_is_valid(
-        data: String,
-        validator: &Validator<T::AuthorityId, T::AccountId>,
-        signature: &ecdsa::Signature,
-    ) -> bool {
-        // verify that the incoming (unverified) pubkey is actually a validator
-        if !Self::is_validator(&validator.account_id) {
-            log::warn!("✋ Account: {:?} is not an authority.", &validator.account_id);
-            return false
-        }
-        let recovered_public_key = recover_public_key_from_ecdsa_signature(signature, &data);
-        if recovered_public_key.is_err() {
-            log::error!(
-                "❌ Recovery of public key from ECDSA Signature: {:?} and data: {:?} failed",
-                &signature,
-                data
-            );
-            return false
-        }
-
-        match T::EthereumPublicKeyChecker::get_validator_for_eth_public_key(
-            &recovered_public_key.expect("Checked for error"),
-        ) {
-            Some(maybe_validator) => maybe_validator == validator.account_id,
-            _ => {
-                log::error!(
-                    "❌ ECDSA signature validation failed on data {:?} validator: {:?} signature {:?}.",
-                    &data,
-                    validator,
-                    signature
-                );
-                false
-            },
-        }
     }
 
     pub fn convert_block_number_to_u32(block_number: BlockNumberFor<T>) -> Result<u32, Error<T>> {
