+++ pallets/parachain-staking/src/lib.rs
@@ -2446,7 +2446,7 @@ pub mod pallet {
             )
             .map_err(|_| DispatchError::Other(Error::<T>::ErrorConvertingBalance.into()))?;
 
-            let function_name: &[u8] = BridgeContractMethod::TriggerGrowth.as_bytes();
+            let function_name: &[u8] = BridgeContractMethod::TriggerGrowth.name_as_bytes();
             let params = vec![
                 (b"uint128".to_vec(), format!("{}", rewards_in_period_128).as_bytes().to_vec()),
                 (
